+++ tests/Unit/Colors/Hsl/ColorTest.php
@@ -12,6 +12,7 @@ use Intervention\Image\Colors\Hsl\Channels\Saturation;
 use Intervention\Image\Colors\Hsl\Color;
 use Intervention\Image\Colors\Hsl\Colorspace;
 use Intervention\Image\Exceptions\NotSupportedException;
+use Intervention\Image\Interfaces\ColorChannelInterface;
 use Intervention\Image\Tests\BaseTestCase;
 
 #[CoversClass(Color::class)]
@@ -82,7 +83,13 @@ final class ColorTest extends BaseTestCase
     public function testNormalize(): void
     {
         $color = new Color(180, 50, 25);
-        $this->assertEquals([.5, 0.5, 0.25, 1], $color->normalizedChannelValues());
+        $this->assertEquals(
+            [.5, 0.5, 0.25, 1],
+            array_map(
+                fn(ColorChannelInterface $channel): float => $channel->normalizedValue(),
+                $color->channels(),
+            )
+        );
     }
 
     public function testToString(): void